Chore extract findByProp to lib/database
