+++ src/components/agent-repo.js
@@ -1,4 +1,4 @@
-const { db, findAll, findById, store } = require('lib/database')
+const { findAll, findById, findByProp, store } = require('lib/database')
 
 const table = 'agents'
 
@@ -7,18 +7,4 @@ module.exports = {
   findById: findById(table),
   findByProp: findByProp(table),
   store: store(table)
-}
-
-function findByProp (table) {
-  return async function findByProp (props) {
-    return await db
-      .select()
-      .table(table)
-      .where(builder =>
-        Object.entries(props)
-          .forEach(([key, value]) => builder.where(key, value))
-      )
-      .limit(1)
-      .then(results => results[0])
-  }
 }
