+++ src/components/project-repo.js
@@ -1,4 +1,4 @@
-const { db, findById, store } = require('lib/database')
+const { findById, findByProp, store } = require('lib/database')
 
 const table = 'projects'
 
@@ -6,18 +6,4 @@ module.exports = {
   findById: findById(table),
   findByProp: findByProp(table),
   store: store(table)
-}
-
-function findByProp (table) {
-  return async function findByProp (props) {
-    return await db
-      .select()
-      .table(table)
-      .where(builder =>
-        Object.entries(props)
-          .forEach(([key, value]) => builder.where(key, value))
-      )
-      .limit(1)
-      .then(results => results[0])
-  }
 }
